backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
apscheduler==3.10.4
httpx==0.27.2
python-multipart==0.0.12
